Explorer tool v 1.0 with input checks

+++ factor_model/risk_calculations/simple_risk_calculation.py
@@ -2,6 +2,6 @@ import pandas as pd
 
 
 def create_ewma_std_estimates(
-    df: pd.DataFrame, halflife: float = 365, min_periods: float = 20
+    df: pd.DataFrame, halflife: float = 365, min_periods: int = 20
 ) -> pd.DataFrame:
     return df.ewm(halflife=halflife, min_periods=min_periods).std().dropna()
